Move middleware to own folder
