+++ src/middleware/render-props.js
